+++ lib/image.ex
@@ -1104,12 +1104,14 @@ defmodule Image do
   * `image` is a a map returned from `Kino.Input.read(image)`
     via a `Kino.Input.image/1` input field. The data will have
     the following fields:
-    * `:file_ref` which contains a file reference to the image. It can be extraced into a path with `Kino.Input.file_path/1`
+
+    * `:file_ref` which contains a file reference to the image. It can be extraced
+      into a path with `Kino.Input.file_path/1`
     * `:width` which is the width of the image in pixels
     * `:height` which is the height of the image in pixels
-    * `:format` which is the image band format which must be `:rgb`
+    * `:format` which is the image band format which must be `:rgb`.
 
-  * `options` is a keyword list of options
+  * `options` is a keyword list of options.
 
   ### Options
 
@@ -1117,6 +1119,9 @@ defmodule Image do
     the image. The default is `3`.
 
   ### Notes
+
+  * Thus function requries Kino v0.11.0 which in turn requires
+    Livebook v0.11.0.
 
   * The image is assumed to contain pixel data that is in
     unsigned 8-bit format which is common for most web-oriented
@@ -1166,6 +1171,9 @@ defmodule Image do
     the image. The default is `3`.
 
   ### Notes
+
+  * Thus function requries Kino v0.11.0 which in turn requires
+    Livebook v0.11.0.
 
   * The image is assumed to contain pixel data that is in
     unsigned 8-bit format which is common for most web-oriented